corrigiendo ARGV.to_f

+++ escape.rb
@@ -1,5 +1,5 @@
 g = ARGV[0].to_f
-r = ARGV[1].to_i
+r = ARGV[1].to_f
 
 v=Math.sqrt(2*g*r)
